adding --all tag to delete/remove eo-Objects

+++ eoxserver/resources/coverages/management/commands/collection.py
@@ -46,11 +46,11 @@ class Command(CommandOutputMixIn, SubParserMixIn, BaseCommand):
         purge_parser = self.add_subparser(parser, 'purge')
         summary_parser = self.add_subparser(parser, 'summary')
         parsers = [
-            create_parser, delete_parser, insert_parser, exclude_parser,
+            create_parser, insert_parser, exclude_parser,
             purge_parser, summary_parser
         ]
 
-        # identifier is a common argument
+        # identifier is a common argument (except for delete it is optional - if --all is tagged -)
         for parser in parsers:
             parser.add_argument(
                 'identifier', nargs=1, help='The collection identifier'
@@ -72,7 +72,17 @@ class Command(CommandOutputMixIn, SubParserMixIn, BaseCommand):
                 '"platform".'
             )
         )
+        delete_parser.add_argument(
+            '--all', '-a', action="store_true",
+            default=False, dest='all_collections',
+            help='When this flag is set, all the collections are selected to be derigesterd'
+        )
 
+
+        delete_parser.add_argument(
+                'identifier', default=None, nargs='?',
+                help='The identifier of the collection to delete.'
+        )
         # common arguments for insertion/exclusion
         insert_parser.add_argument(
             'object_identifiers', nargs='+',
@@ -115,19 +125,18 @@ class Command(CommandOutputMixIn, SubParserMixIn, BaseCommand):
     def handle(self, subcommand, identifier, *args, **kwargs):
         """ Dispatch sub-commands: create, delete, insert, exclude, purge.
         """
-        identifier = identifier[0]
         if subcommand == "create":
-            self.handle_create(identifier, *args, **kwargs)
+            self.handle_create(identifier[0], *args, **kwargs)
         elif subcommand == "delete":
             self.handle_delete(identifier, *args, **kwargs)
         elif subcommand == "insert":
-            self.handle_insert(identifier, *args, **kwargs)
+            self.handle_insert(identifier[0], *args, **kwargs)
         elif subcommand == "exclude":
-            self.handle_exclude(identifier, *args, **kwargs)
+            self.handle_exclude(identifier[0], *args, **kwargs)
         elif subcommand == "purge":
-            self.handle_purge(identifier, *args, **kwargs)
+            self.handle_purge(identifier[0], *args, **kwargs)
         elif subcommand == "summary":
-            self.handle_summary(identifier, *args, **kwargs)
+            self.handle_summary(identifier[0], *args, **kwargs)
 
     def handle_create(self, identifier, type_name, grid_name, **kwargs):
         """ Handle the creation of a new collection.
@@ -158,12 +167,23 @@ class Command(CommandOutputMixIn, SubParserMixIn, BaseCommand):
 
         print('Successfully created collection %r' % identifier)
 
-    def handle_delete(self, identifier, **kwargs):
+    def handle_delete(self, identifier, all_collections,*args, **kwargs):
         """ Handle the deletion of a collection
         """
-        collection = self.get_collection(identifier)
-        collection.delete()
-        print('Successfully deleted collection %r' % identifier)
+        if not all_collections and not identifier:
+            raise CommandError('please specify a collection/s to remove')
+        else:
+            if all_collections:
+                collections = models.Collection.objects.all()
+            elif identifier:
+                collections = [self.get_collection(identifier)]
+            for collection in collections:
+                try:
+                    collection_id = collection.identifier
+                    collection.delete()
+                    self.print_msg('Successfully deregistered collection %r' % collection_id)
+                except models.Collection.DoesNotExist:
+                    raise CommandError('No such Collection %r' % identifier)    
 
     def handle_insert(self, identifier, object_identifiers, **kwargs):
         """ Handle the insertion of arbitrary objects into a collection
